chore: move social media from left panel into footer
