+++ web/containers/Layout/BottomBar/index.tsx
@@ -1,5 +1,14 @@
-import { Badge, Button } from '@janhq/uikit'
+import {
+  Badge,
+  Button,
+  Tooltip,
+  TooltipArrow,
+  TooltipContent,
+  TooltipTrigger,
+} from '@janhq/uikit'
 import { useAtomValue } from 'jotai'
+
+import { FaGithub, FaDiscord } from 'react-icons/fa'
 
 import DownloadingState from '@/containers/Layout/BottomBar/DownloadingState'
 
@@ -26,6 +35,19 @@ const BottomBar = () => {
   const { downloadedModels } = useGetDownloadedModels()
   const { setMainViewState } = useMainViewState()
   const { downloadStates } = useDownloadState()
+
+  const linksMenu = [
+    {
+      name: 'Discord',
+      icon: <FaDiscord size={20} className="flex-shrink-0" />,
+      link: 'https://discord.gg/FTk2MvZwJH',
+    },
+    {
+      name: 'Github',
+      icon: <FaGithub size={16} className="flex-shrink-0" />,
+      link: 'https://github.com/janhq/jan',
+    },
+  ]
 
   return (
     <div className="fixed bottom-0 left-16 z-20 flex h-12 w-[calc(100%-64px)] items-center justify-between border-t border-border bg-background/80 px-3">
@@ -69,11 +91,39 @@ const BottomBar = () => {
 
         <DownloadingState />
       </div>
-      <div className="flex gap-x-2">
-        <SystemItem name="CPU:" value={`${cpu}%`} />
-        <SystemItem name="Mem:" value={`${ram}%`} />
+      <div className="flex items-center gap-x-4">
+        <div className="flex items-center gap-x-2">
+          <SystemItem name="CPU:" value={`${cpu}%`} />
+          <SystemItem name="Mem:" value={`${ram}%`} />
+        </div>
         {/* VERSION is defined by webpack, please see next.config.js */}
-        <span className="text-xs font-semibold ">Jan v{VERSION ?? ''}</span>
+        <span className="text-xs">Jan v{VERSION ?? ''}</span>
+        <div className="mt-1 flex items-center gap-x-2">
+          {linksMenu
+            .filter((link) => !!link)
+            .map((link, i) => {
+              return (
+                <div className="relative" key={i}>
+                  <Tooltip>
+                    <TooltipTrigger>
+                      <a
+                        href={link.link}
+                        target="_blank"
+                        rel="noopener noreferrer"
+                        className="relative flex w-full flex-shrink-0 cursor-pointer items-center justify-center"
+                      >
+                        {link.icon}
+                      </a>
+                    </TooltipTrigger>
+                    <TooltipContent side="top" sideOffset={10}>
+                      <span>{link.name}</span>
+                      <TooltipArrow />
+                    </TooltipContent>
+                  </Tooltip>
+                </div>
+              )
+            })}
+        </div>
       </div>
     </div>
   )
